Move TimerApp composable to app-shared module
