+++ app-shared/src/commonMain/composeResources/values/strings.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<resources>
+    <string name="acc_navigate_back">Go Back</string>
+    <string name="app_name">Timer</string>
+</resources>
